use new interface model
