+++ warp/messages/validator_uptime.go
@@ -4,10 +4,14 @@
 package messages
 
 import (
+	"errors"
 	"fmt"
 
 	"github.com/ava-labs/avalanchego/ids"
+	"github.com/ethereum/go-ethereum/log"
 )
+
+var errInvalidRequest = errors.New("invalid request")
 
 // ValidatorUptime is signed when the ValidationID is known and the validator
 // has been up for TotalUptime seconds.
@@ -48,4 +52,19 @@ func (b *ValidatorUptime) Bytes() []byte {
 
 func (b *ValidatorUptime) initialize(bytes []byte) {
 	b.bytes = bytes
+}
+
+// VerifyMesssage returns nil if the message is valid.
+func (b *ValidatorUptime) VerifyMesssage(sourceAddress []byte) error {
+	// DO NOT USE THIS CODE AS IS IN PRODUCTION.
+
+	// TODO: Does nil/empty SourceAddress matter?
+	if len(sourceAddress) != 0 {
+		return errInvalidRequest
+	}
+
+	log.Info("Received validator uptime message", "validationID", b.ValidationID, "totalUptime", b.TotalUptime)
+	log.Warn("Signing validator uptime message by default, not production behavior", "validationID", b.ValidationID, "totalUptime", b.TotalUptime)
+
+	return nil
 }
